requirements/dev.txt
# Include base requirements
-r base.txt

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
freezegun>=1.2.0
responses>=0.24.0

# Linting & Formatting
ruff>=0.1.0

# Development utilities
ipython>=8.0.0
